+++ sdk/web-pubsub/web-pubsub/test/hubs.spec.ts
@@ -310,7 +310,7 @@ describe("HubClient", function () {
       const res = await client.getClientAccessToken({
         userId: "brian",
         groups: ["group1"],
-        clientType: "mqtt",
+        webPubSubClientAccess: "mqtt",
       });
       const url = new URL(res.url);
       assert.ok(url.searchParams.has("access_token"));